Move permission view itests to a separate file
These tests take a while to run, so it's nice to be able to iterate
on just the relevant tests.
